Move role middleware tests and remove aliasing
