--- coredata_alias.go
@@ -1,6 +1,0 @@
-package goa4web
-
-import common "github.com/arran4/goa4web/core/common"
-
-// CoreData exposes the common CoreData type at the top level package.
-type CoreData = common.CoreData
