+++ internal/router/roles_test.go
@@ -1,23 +1,24 @@
-package goa4web
+package router
 
 import (
 	"context"
-	"github.com/arran4/goa4web/handlers/common"
-	"github.com/arran4/goa4web/internal/middleware"
-	routerpkg "github.com/arran4/goa4web/internal/router"
 	"net/http"
 	"net/http/httptest"
 	"testing"
+
+	corecommon "github.com/arran4/goa4web/core/common"
+	hcommon "github.com/arran4/goa4web/handlers/common"
+	"github.com/arran4/goa4web/internal/middleware"
 )
 
 func TestRoleCheckerMiddlewareAllowed(t *testing.T) {
 	req := httptest.NewRequest("GET", "/admin", nil)
-	ctx := context.WithValue(req.Context(), common.KeyCoreData, &CoreData{SecurityLevel: "administrator"})
+	ctx := context.WithValue(req.Context(), hcommon.KeyCoreData, &corecommon.CoreData{SecurityLevel: "administrator"})
 	req = req.WithContext(ctx)
 
 	called := false
 	h := middleware.NewMiddlewareChain(
-		routerpkg.RoleCheckerMiddleware("administrator"),
+		RoleCheckerMiddleware("administrator"),
 	).Wrap(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		called = true
 		w.WriteHeader(http.StatusOK)
@@ -35,12 +36,12 @@ func TestRoleCheckerMiddlewareAllowed(t *testing.T) {
 
 func TestRoleCheckerMiddlewareDenied(t *testing.T) {
 	req := httptest.NewRequest("GET", "/admin", nil)
-	ctx := context.WithValue(req.Context(), common.KeyCoreData, &CoreData{SecurityLevel: "reader"})
+	ctx := context.WithValue(req.Context(), hcommon.KeyCoreData, &corecommon.CoreData{SecurityLevel: "reader"})
 	req = req.WithContext(ctx)
 
 	called := false
 	h := middleware.NewMiddlewareChain(
-		routerpkg.RoleCheckerMiddleware("administrator"),
+		RoleCheckerMiddleware("administrator"),
 	).Wrap(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		called = true
 	}))
